requirements.txt
pycryptodome>=3.9
python-dotenv>=0.21
pytest>=7.0
